requirements.txt
nltk
tqdm
numpy
scipy
faiss-cpu
